requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.1
confluent-kafka==2.6.0
contourpy==1.3.1
cycler==0.12.1
diagrams==0.24.1
fonttools==4.55.0
graphviz==0.20.3
idna==3.10
influxdb==5.3.2
influxdb-client==1.46.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
msgpack==1.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prometheus_client==0.21.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
reactivex==4.0.4
requests==2.32.3
setuptools==75.2.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Flask==3.1.0
